refactor(profile): 抽取 syncCompanionPlugins 到 profile/companion-sync.js
main.js 最复杂的单函数(232 行,约 28 个依赖)迁出为独立模块:

- syncCompanionPlugins: 配套插件/皮肤同步进桌面专属 profile + 维护
  cordis.patch.yml overlay 行(幂等,已有行不重写;启动/服务重启/
  agent 更新后重放,失败仅记录)
- 主体只引用稳定函数/常量(无可变状态),全部经 createCompanionSync
  依赖注入;__dirname 上溯一层修正(profile/ → ../assets、../builtin-collision)
- removeMarketDuplicate 改为可注入依赖(默认仍用 builtin-collision
  真实实现,运行时行为不变)—— 测试可桩化,顺带消除硬 require
- main.js 净减约 230 行(4557 → 3671,累计 -886)
- electron-builder.yml / check-syntax.js 清单同步
- 新增 4 个组件测试(桩依赖 + 真实临时目录:patch 行生成与二次运行
  幂等、用户移除跳过、无效源跳过、皮肤 ui-skin-* 登记)
- patch-row-heal 的 main.js 接线 pin 测试改指向 companion-sync.js

测试: 456 pass / 1 skip / 0 fail,git diff --check 通过
